Remove bad e10s code from 'sdk/context-menu' (rollback bug#1060138)
