+++ toolkit/jetpack/sdk/context-menu.js
@@ -19,19 +19,14 @@ const { validateOptions, getTypeOf } = require("./deprecated/api-utils");
 const { URL, isValidURI } = require("./url");
 const { WindowTracker, browserWindowIterator } = require("./deprecated/window-utils");
 const { isBrowser, getInnerId } = require("./window/utils");
-const { Ci, Cc, Cu } = require("chrome");
+const { Ci } = require("chrome");
 const { MatchPattern } = require("./util/match-pattern");
+const { Worker } = require("./deprecated/sync-worker");
 const { EventTarget } = require("./event/target");
 const { emit } = require('./event/core');
 const { when } = require('./system/unload');
+const selection = require('./selection');
 const { contract: loaderContract } = require('./content/loader');
-const { omit } = require('./util/object');
-const self = require('./self')
-
-// null-out cycles in .modules to make @loader/options JSONable
-const ADDON = omit(require('@loader/options'), ['modules', 'globals']);
-
-require('../framescript/FrameScriptManager.jsm').enableCMEvents();
 
 // All user items we add have this class.
 const ITEM_CLASS = "addon-context-menu-item";
@@ -64,12 +59,29 @@ const OVERFLOW_MENU_CLASS = "addon-content-menu-overflow-menu";
 // The class of the overflow submenu's xul:menupopup.
 const OVERFLOW_POPUP_CLASS = "addon-content-menu-overflow-popup";
 
+//These are used by PageContext.isCurrent below. If the popupNode or any of
+//its ancestors is one of these, Firefox uses a tailored context menu, and so
+//the page context doesn't apply.
+const NON_PAGE_CONTEXT_ELTS = [
+  Ci.nsIDOMHTMLAnchorElement,
+  Ci.nsIDOMHTMLAppletElement,
+  Ci.nsIDOMHTMLAreaElement,
+  Ci.nsIDOMHTMLButtonElement,
+  Ci.nsIDOMHTMLCanvasElement,
+  Ci.nsIDOMHTMLEmbedElement,
+  Ci.nsIDOMHTMLImageElement,
+  Ci.nsIDOMHTMLInputElement,
+  Ci.nsIDOMHTMLMapElement,
+  Ci.nsIDOMHTMLMediaElement,
+  Ci.nsIDOMHTMLMenuElement,
+  Ci.nsIDOMHTMLObjectElement,
+  Ci.nsIDOMHTMLOptionElement,
+  Ci.nsIDOMHTMLSelectElement,
+  Ci.nsIDOMHTMLTextAreaElement,
+];
+
 // Holds private properties for API objects
 let internal = ns();
-
-function uuid() {
-  return require('./util/uuid').uuid().toString();
-}
 
 function getScheme(spec) {
   try {
@@ -80,22 +92,15 @@ function getScheme(spec) {
   }
 }
 
-let MessageManager = Cc["@mozilla.org/globalmessagemanager;1"].
-                     getService(Ci.nsIMessageBroadcaster);
-
 let Context = Class({
-  initialize: function() {
-    internal(this).id = uuid();
-  },
-
   // Returns the node that made this context current
   adjustPopupNode: function adjustPopupNode(popupNode) {
     return popupNode;
   },
 
   // Returns whether this context is current for the current node
-  isCurrent: function isCurrent(state) {
-    return state;
+  isCurrent: function isCurrent(popupNode) {
+    return false;
   }
 });
 
@@ -104,12 +109,21 @@ let Context = Class({
 let PageContext = Class({
   extends: Context,
 
-  serialize: function() {
-    return {
-      id: internal(this).id,
-      type: "PageContext",
-      args: []
+  isCurrent: function isCurrent(popupNode) {
+    // If there is a selection in the window then this context does not match
+    if (!popupNode.ownerDocument.defaultView.getSelection().isCollapsed)
+      return false;
+
+    // If the clicked node or any of its ancestors is one of the blacklisted
+    // NON_PAGE_CONTEXT_ELTS then this context does not match
+    while (!(popupNode instanceof Ci.nsIDOMDocument)) {
+      if (NON_PAGE_CONTEXT_ELTS.some(function(type) popupNode instanceof type))
+        return false;
+
+      popupNode = popupNode.parentNode;
     }
+
+    return true;
   }
 });
 exports.PageContext = PageContext;
@@ -118,11 +132,19 @@ exports.PageContext = PageContext;
 let SelectionContext = Class({
   extends: Context,
 
-  serialize: function() {
-    return {
-      id: internal(this).id,
-      type: "SelectionContext",
-      args: []
+  isCurrent: function isCurrent(popupNode) {
+    if (!popupNode.ownerDocument.defaultView.getSelection().isCollapsed)
+      return true;
+
+    try {
+      // The node may be a text box which has selectionStart and selectionEnd
+      // properties. If not this will throw.
+      let { selectionStart, selectionEnd } = popupNode;
+      return !isNaN(selectionStart) && !isNaN(selectionEnd) &&
+             selectionStart !== selectionEnd;
+    }
+    catch (e) {
+      return false;
     }
   }
 });
@@ -134,7 +156,6 @@ let SelectorContext = Class({
   extends: Context,
 
   initialize: function initialize(selector) {
-    Context.prototype.initialize.call(this);
     let options = validateOptions({ selector: selector }, {
       selector: {
         is: ["string"],
@@ -144,12 +165,21 @@ let SelectorContext = Class({
     internal(this).selector = options.selector;
   },
 
-  serialize: function() {
-    return {
-      id: internal(this).id,
-      type: "SelectorContext",
-      args: [internal(this).selector]
+  adjustPopupNode: function adjustPopupNode(popupNode) {
+    let selector = internal(this).selector;
+
+    while (!(popupNode instanceof Ci.nsIDOMDocument)) {
+      if (popupNode.mozMatchesSelector(selector))
+        return popupNode;
+
+      popupNode = popupNode.parentNode;
     }
+
+    return null;
+  },
+
+  isCurrent: function isCurrent(popupNode) {
+    return !!this.adjustPopupNode(popupNode);
   }
 });
 exports.SelectorContext = SelectorContext;
@@ -159,7 +189,6 @@ let URLContext = Class({
   extends: Context,
 
   initialize: function initialize(patterns) {
-    Context.prototype.initialize.call(this);
     patterns = Array.isArray(patterns) ? patterns : [patterns];
 
     try {
@@ -169,18 +198,12 @@ let URLContext = Class({
       throw new Error("Patterns must be a string, regexp or an array of " +
                       "strings or regexps: " + err);
     }
+
   },
 
-  isCurrent: function isCurrent(url) {
+  isCurrent: function isCurrent(popupNode) {
+    let url = popupNode.ownerDocument.URL;
     return internal(this).patterns.some(function (p) p.test(url));
-  },
-
-  serialize: function() {
-    return {
-      id: internal(this).id,
-      type: "URLContext",
-      args: []
-    }
   }
 });
 exports.URLContext = URLContext;
@@ -190,7 +213,6 @@ let PredicateContext = Class({
   extends: Context,
 
   initialize: function initialize(predicate) {
-    Context.prototype.initialize.call(this);
     let options = validateOptions({ predicate: predicate }, {
       predicate: {
         is: ["function"],
@@ -200,19 +222,55 @@ let PredicateContext = Class({
     internal(this).predicate = options.predicate;
   },
 
-  isCurrent: function isCurrent(state) {
-    return internal(this).predicate(state);
-  },
-
-  serialize: function() {
-    return {
-      id: internal(this).id,
-      type: "PredicateContext",
-      args: []
-    }
+  isCurrent: function isCurrent(popupNode) {
+    return internal(this).predicate(populateCallbackNodeData(popupNode));
   }
 });
 exports.PredicateContext = PredicateContext;
+
+// List all editable types of inputs.  Or is it better to have a list
+// of non-editable inputs?
+let editableInputs = {
+  email: true,
+  number: true,
+  password: true,
+  search: true,
+  tel: true,
+  text: true,
+  textarea: true,
+  url: true
+};
+
+function populateCallbackNodeData(node) {
+  let window = node.ownerDocument.defaultView;
+  let data = {};
+
+  data.documentType = node.ownerDocument.contentType;
+
+  data.documentURL = node.ownerDocument.location.href;
+  data.targetName = node.nodeName.toLowerCase();
+  data.targetID = node.id || null ;
+
+  if ((data.targetName === 'input' && editableInputs[node.type]) ||
+      data.targetName === 'textarea') {
+    data.isEditable = !node.readOnly && !node.disabled;
+  }
+  else {
+    data.isEditable = node.isContentEditable;
+  }
+
+  data.selectionText = selection.text;
+
+  data.srcURL = node.src || null;
+  data.value = node.value || null;
+
+  while (!data.linkURL && node) {
+    data.linkURL = node.href || null;
+    node = node.parentNode;
+  }
+
+  return data;
+}
 
 function removeItemFromArray(array, item) {
   return array.filter(function(i) i !== item);
@@ -304,82 +362,130 @@ let menuRules = mix(labelledItemRules, {
   }
 });
 
+let ContextWorker = Class({
+  implements: [ Worker ],
+
+  // Calls the context workers context listeners and returns the first result
+  // that is either a string or a value that evaluates to true. If all of the
+  // listeners returned false then returns false. If there are no listeners,
+  // returns true (show the menu item by default).
+  getMatchedContext: function getCurrentContexts(popupNode) {
+    let results = this.getSandbox().emitSync("context", popupNode);
+    if (!results.length)
+      return true;
+    return results.reduce((val, result) => val || result);
+  },
+
+  // Emits a click event in the worker's port. popupNode is the node that was
+  // context-clicked, and clickedItemData is the data of the item that was
+  // clicked.
+  fireClick: function fireClick(popupNode, clickedItemData) {
+    this.getSandbox().emitSync("click", popupNode, clickedItemData);
+  }
+});
+
 // Returns true if any contexts match. If there are no contexts then a
 // PageContext is tested instead
-function hasMatchingContext(contexts, addonInfo) {
-  for (let context of contexts) {
-    if (!(internal(context).id in addonInfo.contextStates)) {
-      console.error("Missing state for context " + internal(context).id + " this is an error in the SDK modules.");
-      return false;
-    }
-    if (!context.isCurrent(addonInfo.contextStates[internal(context).id]))
+function hasMatchingContext(contexts, popupNode) {
+  for (let context in contexts) {
+    if (!context.isCurrent(popupNode))
       return false;
   }
 
   return true;
 }
 
+// Gets the matched context from any worker for this item. If there is no worker
+// or no matched context then returns false.
+function getCurrentWorkerContext(item, popupNode) {
+  let worker = getItemWorkerForWindow(item, popupNode.ownerDocument.defaultView);
+  if (!worker)
+    return true;
+  return worker.getMatchedContext(popupNode);
+}
+
 // Tests whether an item should be visible or not based on its contexts and
 // content scripts
-function isItemVisible(item, addonInfo, usePageWorker) {
+function isItemVisible(item, popupNode, defaultVisibility) {
   if (!item.context.length) {
-    if (!addonInfo.hasWorker)
-      return usePageWorker ? addonInfo.pageContext : true;
+    let worker = getItemWorkerForWindow(item, popupNode.ownerDocument.defaultView);
+    if (!worker)
+      return defaultVisibility;
   }
 
-  if (!hasMatchingContext(item.context, addonInfo))
+  if (!hasMatchingContext(item.context, popupNode))
     return false;
 
-  let context = addonInfo.workerContext;
+  let context = getCurrentWorkerContext(item, popupNode);
   if (typeof(context) === "string" && context != "")
     item.label = context;
 
   return !!context;
 }
 
-// Called when an item is clicked to send out click events to the content
-// scripts
-function itemActivated(item, clickedNode) {
-  let data = {
-    items: [internal(item).id],
-    data: item.data,
-  }
+// Gets the item's content script worker for a window, creating one if necessary
+// Once created it will be automatically destroyed when the window unloads.
+// If there is not content scripts for the item then null will be returned.
+function getItemWorkerForWindow(item, window) {
+  if (!item.contentScript && !item.contentScriptFile)
+    return null;
 
-  while (item.parentMenu) {
-    item = item.parentMenu;
-    data.items.push(internal(item).id);
-  }
+  let id = getInnerId(window);
+  let worker = internal(item).workerMap.get(id);
 
-  let menuData = clickedNode.ownerDocument.defaultView.gContextMenuContentData;
-  let messageManager = menuData.browser.messageManager;
-  messageManager.sendAsyncMessage('sdk/contextmenu/activateitems', data, {
-    popupNode: menuData.popupNode
-  });
-}
+  if (worker)
+    return worker;
 
-function serializeItem(item) {
-  return {
-    id: internal(item).id,
-    contexts: [c.serialize() for (c of item.context)],
+  worker = ContextWorker({
+    window: window,
     contentScript: item.contentScript,
     contentScriptFile: item.contentScriptFile,
-  };
+    onMessage: function(msg) {
+      emit(item, "message", msg);
+    },
+    onDetach: function() {
+      internal(item).workerMap.delete(id);
+    }
+  });
+
+  internal(item).workerMap.set(id, worker);
+
+  return worker;
+}
+
+// Called when an item is clicked to send out click events to the content
+// scripts
+function itemActivated(item, clickedItem, popupNode) {
+  let worker = getItemWorkerForWindow(item, popupNode.ownerDocument.defaultView);
+
+  if (worker) {
+    let adjustedNode = popupNode;
+    for (let context in item.context)
+        adjustedNode = context.adjustPopupNode(adjustedNode);
+    worker.fireClick(adjustedNode, clickedItem.data);
+  }
+
+  if (item.parentMenu)
+    itemActivated(item.parentMenu, clickedItem, popupNode);
 }
 
 // All things that appear in the context menu extend this
 let BaseItem = Class({
   initialize: function initialize() {
-    internal(this).id = uuid();
+    addCollectionProperty(this, "context");
 
-    internal(this).contexts = [];
+    // Used to cache content script workers and the windows they have been
+    // created for
+    internal(this).workerMap = new Map();
+
     if ("context" in internal(this).options && internal(this).options.context) {
       let contexts = internal(this).options.context;
       if (Array.isArray(contexts)) {
         for (let context of contexts)
-          internal(this).contexts.push(context);
+          this.context.add(context);
       }
       else {
-        internal(this).contexts.push(contexts);
+        this.context.add(contexts);
       }
     }
 
@@ -394,72 +500,21 @@ let BaseItem = Class({
       value: internal(this).options.contentScript
     });
 
-    // Resolve URIs here as tests may have overriden self
-    let files = internal(this).options.contentScriptFile;
-    if (files) {
-      if (!Array.isArray(files))
-        files = [files];
-      files = files.map(self.data.url);
-    }
-    internal(this).options.contentScriptFile = files;
     Object.defineProperty(this, "contentScriptFile", {
       enumerable: true,
       value: internal(this).options.contentScriptFile
     });
-
-    // Notify all frames of this new item
-    sendItems([serializeItem(this)]);
   },
 
   destroy: function destroy() {
-    if (internal(this).destroyed)
-      return;
-
-    // Tell all existing frames that this item has been destroyed
-    MessageManager.broadcastAsyncMessage("sdk/contextmenu/destroyitems", {
-      items: [internal(this).id]
-    });
-
     if (this.parentMenu)
       this.parentMenu.removeItem(this);
-
-    internal(this).destroyed = true;
-  },
-
-  get context() {
-    let contexts = internal(this).contexts.slice(0);
-    contexts.add = (context) => {
-      internal(this).contexts.push(context);
-      // Notify all frames that this item has changed
-      sendItems([serializeItem(this)]);
-    };
-    contexts.remove = (context) => {
-      internal(this).contexts = internal(this).contexts.filter(c => {
-        return c != context;
-      });
-      // Notify all frames that this item has changed
-      sendItems([serializeItem(this)]);
-    };
-    return contexts;
-  },
-
-  set context(val) {
-    internal(this).contexts = val.slice(0);
-    // Notify all frames that this item has changed
-    sendItems([serializeItem(this)]);
   },
 
   get parentMenu() {
     return internal(this).parentMenu;
   },
 });
-
-function workerMessageReceived({ data: { id, args } }) {
-  if (internal(this).id != id)
-    return;
-
-  emit(this, ...args);
-}
 
 // All things that have a label on the context menu extend this
 let LabelledItem = Class({
@@ -469,16 +524,11 @@ let LabelledItem = Class({
   initialize: function initialize(options) {
     BaseItem.prototype.initialize.call(this);
     EventTarget.prototype.initialize.call(this, options);
-
-    internal(this).messageListener = workerMessageReceived.bind(this);
-    MessageManager.addMessageListener('sdk/worker/event', internal(this).messageListener);
   },
 
   destroy: function destroy() {
-    if (internal(this).destroyed)
-      return;
-
-    MessageManager.removeMessageListener('sdk/worker/event', internal(this).messageListener);
+    for (let [,worker] of internal(this).workerMap)
+      worker.destroy();
 
     BaseItem.prototype.destroy.call(this);
   },
@@ -662,46 +712,7 @@ exports.Separator = Separator;
 let contentContextMenu = ItemContainer();
 exports.contentContextMenu = contentContextMenu;
 
-function getContainerItems(container) {
-  let items = [];
-  for (let item of internal(container).children) {
-    items.push(serializeItem(item));
-    if (item instanceof Menu)
-      items = items.concat(getContainerItems(item));
-  }
-  return items;
-}
-
-// Notify all frames of these new or changed items
-function sendItems(items) {
-  MessageManager.broadcastAsyncMessage("sdk/contextmenu/createitems", {
-    items,
-    addon: ADDON,
-  });
-}
-
-// Called when a new frame is created and wants to get the current list of items
-function remoteItemRequest({ target: { messageManager } }) {
-  let items = getContainerItems(contentContextMenu);
-  if (items.length == 0)
-    return;
-
-// Context menu support in e10s via https://bugzilla.mozilla.org/show_bug.cgi?id=1060138
-// led to the TypeError in FF38 - https://bugzilla.mozilla.org/show_bug.cgi?id=1130830
-// In FF39+ context-menu.js was rewritten from scratch with processes, but it's not our way.
-// Let's do extra check it for now, though to rollback #1060138 would be better.
-
-  if (messageManager && messageManager.sendAsyncMessage && typeof messageManager.sendAsyncMessage === "function") {
-    messageManager.sendAsyncMessage("sdk/contextmenu/createitems", {
-      items,
-      addon: ADDON,
-    });
-  }
-}
-MessageManager.addMessageListener('sdk/contextmenu/requestitems', remoteItemRequest);
-
 when(function() {
-  MessageManager.removeMessageListener('sdk/contextmenu/requestitems', remoteItemRequest);
   contentContextMenu.destroy();
 });
 
@@ -789,16 +800,16 @@ let MenuWrapper = Class({
 
   // Recurses through the menu setting the visibility of items. Returns true
   // if any of the items in this menu were visible
-  setVisibility: function setVisibility(menu, addonInfo, usePageWorker) {
+  setVisibility: function setVisibility(menu, popupNode, defaultVisibility) {
     let anyVisible = false;
 
     for (let item of internal(menu).children) {
-      let visible = isItemVisible(item, addonInfo[internal(item).id], usePageWorker);
+      let visible = isItemVisible(item, popupNode, defaultVisibility);
 
       // Recurse through Menus, if none of the sub-items were visible then the
       // menu is hidden too.
       if (visible && (item instanceof Menu))
-        visible = this.setVisibility(item, addonInfo, false);
+        visible = this.setVisibility(item, popupNode, true);
 
       let xulNode = this.getXULNodeForItem(item);
       xulNode.hidden = !visible;
@@ -901,7 +912,7 @@ let MenuWrapper = Class({
         if (event.target !== xulNode)
           return;
 
-        itemActivated(item, xulNode);
+        itemActivated(item, item, self.contextMenu.triggerNode);
       }, false);
     }
 
@@ -1016,14 +1027,8 @@ let MenuWrapper = Class({
         this.populate(this.items);
       }
 
-      let mainWindow = event.target.ownerDocument.defaultView;
-      this.contextMenuContentData = mainWindow.gContextMenuContentData;
-      let addonInfo = this.contextMenuContentData.addonInfo[self.id];
-      if (!addonInfo) {
-        console.warn("No context menu state data was provided.");
-        return;
-      }
-      this.setVisibility(this.items, addonInfo, true);
+      let popupNode = event.target.triggerNode;
+      this.setVisibility(this.items, popupNode, PageContext().isCurrent(popupNode));
     }
     catch (e) {
       console.exception(e);
